integrated Resubmission of Requisition Form

+++ LogicUniversityTeam5/Controllers/ViewPastRequestController.cs
@@ -1,4 +1,6 @@
-﻿using ServiceLayer;
+﻿using LogicUniversityTeam5.IdentityHelper;
+using LogicUniversityTeam5.Models;
+using ServiceLayer;
 using ServiceLayer.DataAccess;
 using System;
 using System.Collections.Generic;
@@ -13,11 +15,13 @@ namespace LogicUniversityTeam5.Controllers
     {
 
         IRequisitionService requisitionService;
+        IClassificationService classificationService;
         StationeryStoreEntities context = StationeryStoreEntities.Instance;
 
-        public ViewPastRequestController(RequisitionService requisitionService)
+        public ViewPastRequestController(RequisitionService requisitionService, ClassificationService cs)
         {
             this.requisitionService = requisitionService;
+            classificationService = cs;
         }
 
         //ViewPastRequest/SearchRequisitionForm/E026
@@ -80,21 +84,53 @@ namespace LogicUniversityTeam5.Controllers
         }
 
         //Go to Screen 2.2.1.2b Resubmit Stationery Request Form
+        //Pavana
+        [HttpGet]
         public ActionResult EditSubmittedStationeryRequestForm(int id)
         {
+            CombinedViewModel combinedViewModel = new CombinedViewModel();
             ServiceLayer.DataAccess.Requisition r = requisitionService.getRequisitionById(id);
-            List<ServiceLayer.DataAccess.RequisitionDetail> rdList = r.RequisitionDetails.ToList();
-            return View("ResumbitStationeryRequestForm", rdList);
+            combinedViewModel.Requisitions = r.RequisitionDetails.ToList();
+            TempData["reqID"] = r.RequisitionID;
+            TempData.Keep("reqID");
+            
+            return View("ResubmitStationeryRequestForm", combinedViewModel);
+        }
+
+        //Pavana
+        [HttpPost]
+        public ActionResult ResubmitStationaryRequestForm(CombinedViewModel model)
+        {
+            int reqId = (int)TempData["reqID"];
+            ServiceLayer.DataAccess.Requisition req = requisitionService.getRequisitionById(reqId);
+            string textBoxValue;
+            for (int i = 0; i < model.Requisitions.Count; i++)
+            {
+
+                textBoxValue = model.Requisitions[i].Quantity.ToString().Trim();
+                if (!(textBoxValue.Equals(null)) && !textBoxValue.Equals(""))
+                {
+                    requisitionService.editRequisitionDetailQty(model.Requisitions[i].RequisitionDetailsID, Convert.ToInt32(textBoxValue));
+                }
+            }
+
+            requisitionService.submitRequisition(req.RequisitionID);
+            return RedirectToAction("EditSubmittedStationeryRequestForm", new { id = reqId, isSubmit = true });
         }
 
         //This is for delete in Screen 2.2.1.2b
+        //Pavana
+        [HttpGet]
         public ActionResult DeleteRequestedItems(int id)
         {
-            requisitionService.deleteRequisitionDetail(id);
+            int reqId = (int)TempData["reqID"];
+            //Delete from existing reqDetails in DB
+            if (id > 0)
+            {
+                requisitionService.deleteRequisitionDetail(id);
+            }
 
-            //get reqId using reqdetailId
-            ServiceLayer.DataAccess.RequisitionDetail rd = context.RequisitionDetails.FirstOrDefault(rq => rq.RequisitionDetailsID == id);
-            return RedirectToAction("EditSubmittedStationeryRequestForm", new { id = rd.Requisition.RequisitionID });
+            return RedirectToAction("EditSubmittedStationeryRequestForm", new { id = reqId});
         }
 
         public ActionResult EmployeeHome()
@@ -102,6 +138,68 @@ namespace LogicUniversityTeam5.Controllers
             return View();
         }
 
+        public ActionResult ViewStationeryCatalogue()
+        {
+            TempData.Keep("reqID");
+            CombinedViewModel combinedView = new CombinedViewModel();
+
+            combinedView.Items = classificationService.GetItems();
+            combinedView.Categories = classificationService.GetCategories();
+            combinedView.categorySelected = "All";
+            combinedView.AddedText = new List<string>();
+            foreach (Item i in combinedView.Items)
+            {
+                combinedView.AddedText.Add(" ");
+            }
+            return View(combinedView);
+        }
+
+        [HttpPost]
+        public ActionResult ViewStationeryCatalogue(CombinedViewModel model)
+        {
+            //getting the logged in user
+            string currentLoggedInEmployeeId = User.Identity.GetEmployeeId();
+
+            //getting requisitionID
+            int reqId = (int)TempData["reqID"];
+            TempData.Keep("reqID");
+
+            //getting req for  the reqId
+            ServiceLayer.DataAccess.Requisition existingReqOfEmployee = requisitionService.getRequisitionById(reqId);
+
+            List<RequisitionDetail> existingReqDetail = requisitionService.getRequisitionDetails(reqId);
+
+            List<string> existingReqDetailsItemIds = existingReqDetail.Select(r => r.ItemID).ToList();
+
+            //Looping each item in view
+            for (int i = 0; i < model.Items.Count; i++)
+            {
+                string textBoxValue = model.AddedText[i].ToString().Trim();
+                string itemIdInView = model.Items[i].ItemID;
+                //if item is existing item in requistion, add to quantity
+                if (existingReqDetailsItemIds.Contains(itemIdInView) && (textBoxValue != null && textBoxValue != ""))
+                {
+                    RequisitionDetail rd =
+                        existingReqOfEmployee.RequisitionDetails.First(x => x.ItemID.Equals(itemIdInView));
+
+                    rd.Quantity += Convert.ToInt32(textBoxValue);
+                    requisitionService.editRequisitionDetailQty(rd.RequisitionDetailsID, rd.Quantity);
+                }
+                //if item is not existing item in requistion, create new reqdetails
+                else if (textBoxValue != null && textBoxValue != "")
+                {
+
+                    RequisitionDetail newRD = new RequisitionDetail();
+                    newRD.RequisitionID = existingReqOfEmployee.RequisitionID;
+                    requisitionService
+                        .addNewRequisitionDetail(existingReqOfEmployee.RequisitionID,
+                                                itemIdInView,
+                                                Convert.ToInt32(textBoxValue));
+                }
+            }
+
+            return RedirectToAction("EditSubmittedStationeryRequestForm", new { id = reqId });
+        }
         /*
         public ActionResult StationeryCatalogueView(int id)
         {
